+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Version 0.7.9
 -------------
 
 * metadata: add ``nohz_full`` and ``isolated`` to ``cpu_config``
+* add ``--affinity`` option to the ``metadata`` command
 
 Version 0.7.8 (2016-09-10)
 --------------------------